Remove syntax-based retrieval
