+++ crates/edit_prediction_context/src/assemble_excerpts.rs
