+++ crates/edit_prediction_context/src/fake_definition_lsp.rs
