--- crates/edit_prediction_context2/Cargo.toml
@@ -1,42 +1,0 @@
-[package]
-name = "edit_prediction_context2"
-version = "0.1.0"
-edition.workspace = true
-publish.workspace = true
-license = "GPL-3.0-or-later"
-
-[lints]
-workspace = true
-
-[lib]
-path = "src/edit_prediction_context2.rs"
-
-[dependencies]
-parking_lot.workspace = true
-anyhow.workspace = true
-collections.workspace = true
-futures.workspace = true
-gpui.workspace = true
-language.workspace = true
-lsp.workspace = true
-project.workspace = true
-log.workspace = true
-serde.workspace = true
-smallvec.workspace = true
-tree-sitter.workspace = true
-util.workspace = true
-
-[dev-dependencies]
-env_logger.workspace = true
-indoc.workspace = true
-futures.workspace = true
-gpui = { workspace = true, features = ["test-support"] }
-language = { workspace = true, features = ["test-support"] }
-lsp = { workspace = true, features = ["test-support"] }
-pretty_assertions.workspace = true
-project = {workspace= true, features = ["test-support"]}
-serde_json.workspace = true
-settings = {workspace= true, features = ["test-support"]}
-text = { workspace = true, features = ["test-support"] }
-util = { workspace = true, features = ["test-support"] }
-zlog.workspace = true
