--- crates/edit_prediction_context2/LICENSE-GPL
@@ -1,1 +1,0 @@
-../../LICENSE-GPL
