+++ crates/zeta/Cargo.toml
@@ -30,7 +30,6 @@ credentials_provider.workspace = true
 db.workspace = true
 edit_prediction.workspace = true
 edit_prediction_context.workspace = true
-edit_prediction_context2.workspace = true
 editor.workspace = true
 feature_flags.workspace = true
 fs.workspace = true
